Fix double serialization
