+++ engine/src/main/kotlin/world/gregs/voidps/engine/data/FileStorage.kt
@@ -2,14 +2,15 @@ package world.gregs.voidps.engine.data
 
 import com.fasterxml.jackson.core.JsonGenerator
 import com.fasterxml.jackson.databind.ObjectMapper
-import com.fasterxml.jackson.databind.SerializerProvider
 import com.fasterxml.jackson.databind.exc.MismatchedInputException
 import com.fasterxml.jackson.databind.module.SimpleModule
-import com.fasterxml.jackson.databind.ser.std.StdSerializer
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory
 import com.fasterxml.jackson.dataformat.yaml.YAMLGenerator
 import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
 import com.github.michaelbull.logging.InlineLogger
+import world.gregs.voidps.engine.data.serial.DoubleArraySerializer
+import world.gregs.voidps.engine.data.serial.DoubleSerializer
+import world.gregs.voidps.engine.data.serial.TileSerializer
 import world.gregs.voidps.engine.map.Tile
 import java.io.File
 
@@ -66,20 +67,8 @@ class FileStorage private constructor(
         private fun jsonMapper() = jacksonObjectMapper().apply {
             enable(JsonGenerator.Feature.WRITE_BIGDECIMAL_AS_PLAIN)
             val module = SimpleModule()
-            module.addSerializer(DoubleArray::class.java, object : StdSerializer<DoubleArray>(DoubleArray::class.java) {
-                override fun serialize(value: DoubleArray, gen: JsonGenerator, provider: SerializerProvider) {
-                    gen.writeStartArray()
-                    for (double in value) {
-                        gen.writeNumber(double.toBigDecimal())
-                    }
-                    gen.writeEndArray()
-                }
-            })
-            module.addSerializer(Double::class.java, object : StdSerializer<Double>(Double::class.java) {
-                override fun serialize(value: Double, gen: JsonGenerator, provider: SerializerProvider) {
-                    gen.writeNumber(value.toBigDecimal())
-                }
-            })
+            module.addSerializer(DoubleArray::class.java, DoubleArraySerializer)
+            module.addSerializer(Double::class.java, DoubleSerializer)
             module.addSerializer(Tile::class.java, TileSerializer)
             registerModule(module)
         }
